MiniTest-OECT_QT_dev_single/requirements.txt
# 单设备OECT瞬态测试上位机 依赖

PyQt5>=5.15.0
pyqtgraph>=0.13.0
pyserial>=3.5
pyserial-asyncio>=0.6
numpy>=1.21.0
